refactor(flight): split the buffered drive loops out of producer.rs (#3095, epic #1116)
The NB2 wiring grew producer.rs 3437 -> 3445, which the file-size ratchet FAILs
(it is 4x over the 800-line source target). Split by responsibility instead of
waiving: new producer_drive.rs holds the two WHOLE-PARTITION drive loops —
drive_merge (the buffered collect route behind produce / produce_from_paths /
produce_from_resolved) and drive_aggregate (the aggregation route) — as the sibling
of producer_stream.rs, which already owns the row-granular streaming loop.

producer.rs 3437 -> 3243 (a net REDUCTION vs main). No behaviour change:
cqlite-flight --lib 383 pass, plus the three static/differential/oracle targets.
